# packages/domain-types/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_TestRunRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TestRunRead"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[TestRunRead]",
        "type": "object"
      },
...
      "TestOutcome": {
        "enum": [
          "PASS",
          "FAIL",
          "INFO"
        ],
        "title": "TestOutcome",
        "type": "string"
      },
...
      "TestRunRead": {
        "properties": {
          "component": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ComponentSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "conditions": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conditions"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "equipment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equipment"
          },
          "experiment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExperimentSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "identifier": {
            "title": "Identifier",
            "type": "string"
          },
          "is_placeholder": {
            "title": "Is Placeholder",
            "type": "boolean"
          },
          "measurements": {
            "items": {
              "$ref": "#/components/schemas/MeasurementRead"
            },
            "title": "Measurements",
            "type": "array"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "outcome": {
            "$ref": "#/components/schemas/TestOutcome"
          },
          "part_instance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PartInstanceSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "performed_at": {
            "format": "date-time",
            "title": "Performed At",
            "type": "string"
          },
          "performed_by": {
            "title": "Performed By",
            "type": "string"
          },
          "prototype": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PrototypeSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "revision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RevisionSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "test_type": {
            "$ref": "#/components/schemas/TestTypeSummary"
          },
          "timing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TimingSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "updated_by": {
            "title": "Updated By",
            "type": "string"
          },
          "watch": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WatchSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "created_at",
          "created_by",
          "updated_at",
          "updated_by",
          "id",
          "identifier",
          "title",
          "test_type",
          "performed_at",
          "performed_by",
          "outcome",
          "is_placeholder",
          "prototype",
          "watch",
          "part_instance",
          "revision",
          "component",
          "experiment",
          "equipment",
          "conditions",
          "notes",
          "measurements",
          "timing"
        ],
        "title": "TestRunRead",
        "type": "object"
      },
...
      "TestTypeCreate": {
        "properties": {
          "allow_custom_metrics": {
            "default": false,
            "title": "Allow Custom Metrics",
            "type": "boolean"
          },
          "code": {
            "maxLength": 32,
            "minLength": 2,
            "pattern": "^[A-Za-z][A-Za-z0-9_]*$",
            "title": "Code",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/MetricDef"
            },
            "title": "Metrics",
            "type": "array"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "uses_positions": {
            "default": false,
            "title": "Uses Positions",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "name"
        ],
        "title": "TestTypeCreate",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nInternal engineering platform for **NEMETH \u2014 Detroit**: product definition, components\nwith immutable revisions, recursive bills of materials, and (in later slices) prototypes,\nexperiments, measurements and serialized watch genealogy.\n\nErrors are returned as RFC 9457 problem details (`application/problem+json`).\nResources are addressed by UUID or by human identifier (`N1-MVT-002`).\n",
    "title": "NEMETH Engineering Platform",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/experiments/{ref}/test-runs": {
      "get": {
        "operationId": "experiment_test_runs_api_v1_experiments__ref__test_runs_get",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TestRunRead"
                  },
                  "title": "Response Experiment Test Runs Api V1 Experiments  Ref  Test Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Experiment Test Runs",
        "tags": [
          "testing"
        ]
      }
    },
...
    "/api/v1/part-instances/{ref}/test-runs": {
      "get": {
        "operationId": "part_instance_test_runs_api_v1_part_instances__ref__test_runs_get",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TestRunRead"
                  },
                  "title": "Response Part Instance Test Runs Api V1 Part Instances  Ref  Test Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Part Instance Test Runs",
        "tags": [
          "testing"
        ]
      }
    },
...
    "/api/v1/prototypes/{ref}/test-runs": {
      "get": {
        "operationId": "prototype_test_runs_api_v1_prototypes__ref__test_runs_get",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TestRunRead"
                  },
                  "title": "Response Prototype Test Runs Api V1 Prototypes  Ref  Test Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prototype Test Runs",
        "tags": [
          "testing"
        ]
      }
    },
    "/api/v1/prototypes/{ref}/timing": {
      "get": {
        "description": "Timing summary of the most recent timegrapher run on this prototype.",
        "operationId": "prototype_timing_api_v1_prototypes__ref__timing_get",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/TimingSummary"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Prototype Timing Api V1 Prototypes  Ref  Timing Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prototype Timing",
        "tags": [
          "testing"
        ]
      }
    },
...
    "/api/v1/test-runs": {
      "get": {
        "operationId": "list_test_runs_api_v1_test_runs_get",
        "parameters": [
          {
            "description": "Test type code",
            "in": "query",
            "name": "test_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Test type code",
              "title": "Test Type"
            }
          },
          {
            "in": "query",
            "name": "prototype_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Prototype Id"
            }
          },
          {
            "in": "query",
            "name": "part_instance_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Part Instance Id"
            }
          },
          {
            "in": "query",
            "name": "experiment_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Experiment Id"
            }
          },
          {
            "in": "query",
            "name": "revision_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Revision Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_TestRunRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Test Runs",
        "tags": [
          "testing"
        ]
      },
      "post": {
        "operationId": "create_test_run_api_v1_test_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestRunCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestRunRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Test Run",
        "tags": [
          "testing"
        ]
      }
    },
    "/api/v1/test-runs/{ref}": {
      "get": {
        "operationId": "get_test_run_api_v1_test_runs__ref__get",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestRunRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Test Run",
        "tags": [
          "testing"
        ]
      },
      "patch": {
        "operationId": "update_test_run_api_v1_test_runs__ref__patch",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestRunUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestRunRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Test Run",
        "tags": [
          "testing"
        ]
      }
    },
    "/api/v1/test-runs/{ref}/measurements": {
      "post": {
        "operationId": "add_measurements_api_v1_test_runs__ref__measurements_post",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/MeasurementCreate"
                },
                "title": "Payload",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestRunRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Measurements",
        "tags": [
          "testing"
        ]
      }
    },
    "/api/v1/test-types": {
      "get": {
        "operationId": "list_test_types_api_v1_test_types_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_inactive",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Inactive",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TestTypeRead"
                  },
                  "title": "Response List Test Types Api V1 Test Types Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Test Types",
        "tags": [
          "testing"
        ]
      },
      "post": {
        "operationId": "create_test_type_api_v1_test_types_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestTypeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestTypeRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Test Type",
        "tags": [
          "testing"
        ]
      }
    },
    "/api/v1/test-types/{ref}": {
      "get": {
        "operationId": "get_test_type_api_v1_test_types__ref__get",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestTypeRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Test Type",
        "tags": [
          "testing"
        ]
      },
      "patch": {
        "operationId": "update_test_type_api_v1_test_types__ref__patch",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestTypeUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestTypeRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Test Type",
        "tags": [
          "testing"
        ]
      }
    },
...
    "/api/v1/watches/{ref}/test-runs": {
      "get": {
        "operationId": "watch_test_runs_api_v1_watches__ref__test_runs_get",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TestRunRead"
                  },
                  "title": "Response Watch Test Runs Api V1 Watches  Ref  Test Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Watch Test Runs",
        "tags": [
          "testing"
        ]
      }
    },
    "/api/v1/watches/{ref}/timing": {
      "get": {
        "operationId": "watch_timing_api_v1_watches__ref__timing_get",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/TimingSummary"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Watch Timing Api V1 Watches  Ref  Timing Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Watch Timing",
        "tags": [
          "testing"
        ]
      }
    }
  }
}